==> scripts/05 - create_combinations.ipynb <==
import pandas as pd

home = list()
home_sub = list()
relatives = list()
rel_region = list()

for i in range(158):
    df = pd.read_csv("../03 - Combined/user"+str(i)+".csv")
    homedf = df[df['where'] == "Home"]
    home.append(homedf['city'].fillna("").mode()[0] if not homedf.empty else "")
    home_sub.append(homedf['suburb'].fillna("").mode()[0] if not homedf.empty else "")
    reldf = df[df['where'] == "Relatives home"]
    relatives.append(reldf['city'].fillna("").mode()[0] if not reldf.empty else "")
    rel_region.append(reldf['region'].fillna("").mode()[0] if not reldf.empty else "")

finaldf = pd.DataFrame([x for x in range (158)], columns=['userid'])
finaldf['home_sub'] = home_sub
finaldf['home_city'] = home
finaldf['relatives_city'] = relatives
finaldf['relatives_region'] = rel_region
finaldf.to_csv("userCities.csv", index=False)

From here, heuristic was used to create some combinations of users which seems to have common characteristics in terms of home location.

all = [x for x in range (158)]
avoid = [8, 12, 17, 36, 38, 49, 54, 64, 65, 72, 75, 79, 90, 103, 116, 138]
listone = []
c1 = [13, 24, 30, 40, 53, 77, 93, 107, 118, 126, 136, 155]
c2 = [7, 21, 22, 48, 56, 66, 70, 86, 102, 104, 144, 151]
c3 = [5, 41, 43, 55, 82, 117, 121, 122, 132, 133, 135, 148]
c4 = [4, 14, 16, 20, 44, 61, 68, 74, 95, 109, 114, 115]
c5 = [2, 6, 9, 23, 45, 52, 87, 106, 125, 141, 142, 153]
c6 = [1, 18, 49, 78, 92, 99, 100, 119, 130, 139, 146, 156]

avoid.extend(c1)
avoid.extend(c2)
avoid.extend(c3)
avoid.extend(c4)
avoid.extend(c5)
avoid.extend(c6)
eliminated = set(avoid)
remaining = set(all) - eliminated
remaining = list(remaining)

import random
random.shuffle(remaining)
listone.append(c1)
listone.append(c2)
listone.append(c3)
listone.append(c4)
listone.append(c5)
listone.append(c6)
while(len(remaining) >= 12):
    extracted = remaining[:12]
    extracted.sort()
    listone.append(extracted)
    remaining = remaining[12:]
listone
